vterm-minibuffer: made it possible to run multiple shells 4 same dir
The prefix arg allows you to specify which of the shells to use.

M-x vterm-minibuffer       => start new vterm shell (id: 1)
C-u 5 M-x vterm-minibuffer => start new vterm shell (id: 5)
C-u 1 M-x vterm-minibuffer => use existing vterm shell with id 1

This number id of each vterm shell is referred to as the subid. These
values are only shared among shells in the same
default-directory. That is to say, you could have a shell with subid 7
in /var and another shell with subid 7 in /home and they wouldn't conflict.